🚚 move json and translation into /data folder 🔥 removed unused json
