+++ src/data/translations.js
