--- src/js/____s2s_data.json
@@ -1,21 +1,0 @@
-[
-	{ 
-		"lat": 50.075296,
-		"lon": 14.405987,
-		"title": "Smíchovská střední průmyslová škola a gymnázium",
-		"text": "Naše škola je od března roku 2022 prvním účastníkem pilotního projektu „senzory do škol“. Jedná se o iniciativu veřejného sběru dat, konkrétně kvality ovzduší. Naše škola již disponuje meteostanicí, kterou si studenti sami postavili a sbírá veličiny v podobě intenzity prachu, teploty, vlhkosti a tlaku.",
-		"contacts_title": "Contact",
-		"contacts":[
-			{ "name": "Alexandr Waage", "url": "mailto:reditel@ssps.cz" }
-		]
-,"partner_title": "Partner",
-"partner":[
-	{ "name": "Senzorvzduchu.cz", "url": "https://www.senzorvzduchu.cz/" }
-]
-,"website": "https://www.ssps.cz/projekty/iot/senzory-do-skol/"
-,"facebook_page": "https://www.facebook.com/smichovskasps"
-	}
-]
-
-
-
